requirements.txt
asn1crypto==0.24.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Flask==1.1.1
future==0.17.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pycparser==2.19
python-telegram-bot==12.1.1
requests==2.22.0
six==1.12.0
tornado==6.0.3
urllib3==1.25.6
Werkzeug==0.16.0
pandas==0.20.3
geopandas==0.3.0
geocoder==1.32.1
googlemaps==2.5.1
geojsonio==0.0.3
Shapely==1.6.1